Replaced the studio page with the updated one
