--- src/pages/studio-new.md
@@ -1,49 +1,0 @@
----
-templateKey: Studio2
-title: Random Studio
-introBlock:
-  - copy: We create brand and service experiences that trigger a sense of wonder.
-    image: /img/wec190502_chanel_j12_022_2.jpg
-    video: /img/boat.mp4
-  - copy: 'Our work comes to life in the space where the digital meets the physical. '
-    image: /img/6bea20f020e723539956196ada8b6b35.v-vvv-shanghai-49-r3-jpg.jpg
-    video: /img/digital.mp4
-services:
-  list:
-    - copy: |
-        Interactive brand and service experiences,
-        including strategy, spatial design and engineering.
-      image: /img/connected-space-02-lv-2x.jpg
-      title: |-
-        Experience Design For
-        Retail and Exhibitions
-    - copy: |-
-        Strategy and continuous development
-        for services, tools and spatial infrastructure.
-      image: /img/studio_expertise2.png
-      title: Product Development
-    - copy: |-
-        R&D Partnerships including bespoke trend reports,
-        future concepts and prototypes.
-      image: /img/studio_expertise3.png
-      title: Design Research
-  title: 'Our Areas of Expertise:'
-skillset:
-  - Strategy
-  - Concept
-  - Art Direction
-  - Spatial Design
-  - UX Design
-  - Engineering
-studioImpression:
-  images:
-    - image: /img/studio1.jpg
-    - image: /img/studio2.jpg
-    - image: /img/studio3.jpg
-    - image: /img/studio4.jpg
-    - image: /img/studio5.jpg
-    - image: /img/studio6.jpg
-    - image: /img/studio7.jpg
-  showIndicator: true
-  title: Studio Impressions
----
