+++ src/pages/studio.md
@@ -1,109 +1,49 @@
 ---
 templateKey: Studio
-intro: >-
-  Random is an independent experience design studio.<br /> We are strategists,
-  designers and engineers from all over the world, collaborating with brands to
-  trigger curiosity and use technology to spark new, unexpected connections
-  between people and the space around them. <br> <br> How can we make the
-  digital a living, breathing part of our world? How can technology be
-  integrated into the very foundations of a space; as part of our physical
-  world, engaging us to explore, play and wonder?
-
-  <br> <br> We have a holistic, research-led approach that can be applied to
-  products, experiences and platforms where spatial and digital design are
-  integrated.
-infoBlock:
-  - collection:
-      - images:
-          - caption: >-
-              [Exhibition for Louis
-              Vuitton](https://random.studio/projects/travel-exhibition-louis-vuitton)
-            image: /img/connected-space-02-lv-2x.jpg
-          - caption: >-
-              [In-store Installations for
-              NikeLab](https://random.studio/projects/NikeLab-ACG)
-            image: /img/connected-space-01-Nike-ACG@2x.jpg
-          - caption: >-
-              [Exhibition for
-              Hermès](https://random.studio/projects/hermes-festival-des-metiers)
-            image: /img/connected-space-03-Hermes@2x.jpg
-        info: ''
-        showIndicator: true
-      - info: >-
-          Connected Spaces
-
-
-          Blending digital and physical space, we create immersive, sensory
-          experiences that open new worlds and evoke a sense of wonder. Our
-          interactive environments are shaped by connectivity: they use
-          technology in a tactile way, inviting people to play, reflect and
-          explore.
-
-
-          * Experiental installations
-
-          * Exhibitions
-
-          * In-store experiences
-        showIndicator: false
-  - collection:
-      - info: >-
-          Human-Centered Products
-
-
-          Drawing on in-depth research and technical expertise, we craft
-          holistic global platforms for businesses and consumers that are both
-          relevant and enduring. Extending beyond their primary function, the
-          digital tools we build are experiential and boundary-pushing: they
-          rethink and adapt to the way brands work today.
-
-
-          * Consumer and B2B platforms
-
-          * Omnichannel services
-        showIndicator: false
-      - images:
-          - caption: Shoppable wall for Tommy Hilfiger
-            image: /img/human-centered-products-02-gigi-shopwall-2x.jpg
-          - caption: Digital showroom for PVH
-            image: /img/human-centered-products-02-Digital-Showroom@2x.jpg
-        showIndicator: true
-  - collection:
-      - images:
-          - caption: '[Website for Raf Simons](https://random.studio/projects/rafsimons)'
-            image: /img/communication-design-01-raf-simons-2x.jpg
-          - caption: Communication strategy for Fred Perry
-            image: /img/communication-design-02-Raf-x-Fred-Perry@2x.jpg
-          - caption: Interactive documentary 51 Sprints
-            image: /img/communication-design-03-51-Sprints@2x.jpg
-        showIndicator: true
-      - info: >-
-          Communication Formats
-
-
-          Sparking new conversations between brands and their audiences, we
-          disrupt existing formats to develop unique ways of communicating. Each
-          project needs its own voice: we play with radical forms that capture
-          the energy and many layers of a story in imaginative and surprising
-          ways.
-
-
-          * Conceptualising new formats
-
-          * Social engagement
-
-          * Film direction and editing
-        showIndicator: false
+title: Random Studio
+introBlock:
+  - copy: We create brand and service experiences that trigger a sense of wonder.
+    image: /img/wec190502_chanel_j12_022_2.jpg
+    video: /img/boat.mp4
+  - copy: 'Our work comes to life in the space where the digital meets the physical. '
+    image: /img/6bea20f020e723539956196ada8b6b35.v-vvv-shanghai-49-r3-jpg.jpg
+    video: /img/digital.mp4
+services:
+  list:
+    - copy: |
+        Interactive brand and service experiences,
+        including strategy, spatial design and engineering.
+      image: /img/connected-space-02-lv-2x.jpg
+      title: |-
+        Experience Design For
+        Retail and Exhibitions
+    - copy: |-
+        Strategy and continuous development
+        for services, tools and spatial infrastructure.
+      image: /img/studio_expertise2.png
+      title: Product Development
+    - copy: |-
+        R&D Partnerships including bespoke trend reports,
+        future concepts and prototypes.
+      image: /img/studio_expertise3.png
+      title: Design Research
+  title: 'Our Areas of Expertise:'
+skillset:
+  - Strategy
+  - Concept
+  - Art Direction
+  - Spatial Design
+  - UX Design
+  - Engineering
 studioImpression:
   images:
-    - image: /img/studio_01.jpg
-    - image: /img/studio_02.jpg
-    - image: /img/studio_03.jpg
-    - image: /img/studio_04.jpg
-    - image: /img/studio_05.jpg
-    - image: /img/studio_06.jpg
-    - image: /img/studio_07.jpg
+    - image: /img/studio1.jpg
+    - image: /img/studio2.jpg
+    - image: /img/studio3.jpg
+    - image: /img/studio4.jpg
+    - image: /img/studio5.jpg
+    - image: /img/studio6.jpg
+    - image: /img/studio7.jpg
   showIndicator: true
   title: Studio Impressions
 ---
-
