--- src/templates/Studio/InfoBlock/InfoBlock.module.scss
@@ -1,120 +1,0 @@
-.infoBlock {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-}
-
-.info {
-  order: 2;
-  margin-bottom: 20vh;
-  padding: 1.4rem;
-
-  p {
-    margin-bottom: 1.2rem;
-  }
-
-  ul {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-
-    li::before {
-      content: "- ";
-    }
-  }
-}
-
-.carouselWrapper {
-  font-size: 1rem;
-}
-
-@media (min-width: 36rem) {
-  .infoBlock {
-    &:nth-of-type(1),
-    &:nth-of-type(2) {
-      .carouselWrapper {
-        width: 60%;
-      }
-    }
-
-    &:nth-of-type(2) {
-      .carouselWrapper {
-        align-self: flex-end;
-
-        p,
-        span {
-          margin: 0;
-        }
-
-        .info {
-          max-width: 48rem;
-          margin: 0 auto;
-        }
-      }
-    }
-
-    &:nth-of-type(3) {
-      .carouselWrapper {
-        width: 100%;
-      }
-    }
-  }
-
-  .info {
-    margin: 0;
-    padding: 4.8rem 2rem 7.2rem;
-
-    p {
-      margin-bottom: 1.8rem;
-    }
-  }
-}
-
-@media (min-width: 72rem) {
-  .infoBlock {
-    flex-direction: row;
-    margin: 0 auto;
-    margin-bottom: 20vh;
-    font-size: 1.74rem;
-    align-items: center;
-
-    &:nth-of-type(1) {
-      .carouselWrapper {
-        min-width: 30%;
-        width: 30%;
-      }
-
-      .info {
-        max-width: 53rem;
-      }
-    }
-
-    &:nth-of-type(2) {
-      max-width: 72rem;
-
-      .carouselWrapper {
-        min-width: 43%;
-        width: 43%;
-      }
-    }
-
-    &:nth-of-type(3) {
-      flex-direction: column;
-      align-items: initial;
-      .carouselWrapper {
-        min-width: 65%;
-        width: 65%;
-      }
-
-      .info {
-        max-width: 72rem;
-        margin: 3.2rem auto 0;
-      }
-    }
-  }
-
-  .info {
-    order: 0;
-    padding: 0 3.2rem;
-  }
-}
