--- src/templates/Studio/Intro/Intro.module.scss
@@ -1,138 +1,0 @@
-@import '../../../styles/mixins';
-
-.logo {
-  @include blendedText;
-  position: relative;
-  font-family: 'Larish Neue', serif;
-  font-weight: normal;
-  font-size: 3.2rem;
-  mix-blend-mode: difference;
-  margin: 0;
-  padding-top: 40vh;
-  padding-left: 1.4rem;
-  letter-spacing: -0.01em;
-  line-height: 1.1;
-}
-
-.intro {
-  position: relative;
-  margin: 15.4rem 0 0;
-  padding: 0 1.4rem;
-}
-.introHeader {
-  margin-bottom: 6.4rem;
-  font-size: var(--font-small);
-  font-weight: normal;
-  letter-spacing: 0.01em;
-}
-
-.address {
-  font-size: 1rem;
-  margin-bottom: 3.2rem;
-  padding-top: 0.3rem;
-  letter-spacing: 0.02em;
-
-  p {
-    margin-bottom: 1.3em;
-  }
-
-  div:last-of-type {
-    p:last-of-type {
-      margin-bottom: 1em;
-    }
-  }
-
-  a {
-    position: relative;
-    display: inline-block;
-    text-decoration: none;
-    &::after {
-      content: '';
-      position: absolute;
-      bottom: 1px;
-      left: 0;
-      height: 1px;
-      width: 100%;
-      background-color: black;
-    }
-  }
-
-  // Remove underline of phonenumber
-  div:last-of-type {
-    a:first-of-type::after {
-      display: none;
-    }
-  }
-}
-
-@media (min-width: 36rem) {
-  .logo {
-    padding-left: 4rem;
-    font-size: 5rem;
-  }
-
-  .intro {
-    padding: 4rem;
-  }
-
-  .introHeader {
-    font-size: var(--font-med);
-    margin-bottom: 9.6rem;
-  }
-
-  .address {
-    float: left;
-    font-size: var(--font-small);
-    margin: 0;
-    margin-right: 2rem;
-
-    a::after {
-      height: 2px;
-    }
-  }
-}
-
-@media (min-width: 54rem) {
-  .logo,
-  .intro {
-    max-width: 48rem;
-    margin: 0 auto;
-  }
-
-  .logo {
-    padding-top: 40vh;
-  }
-
-  .intro {
-    margin-top: 7.2vw;
-  }
-
-  .introHeader {
-    font-size: var(--font-med);
-  }
-
-  .address {
-    padding-top: 0.5rem;
-  }
-}
-
-@media (min-width: 72rem) {
-  .logo,
-  .intro {
-    max-width: 64rem;
-  }
-
-  .logo {
-    max-width: 72rem;
-    font-size: 5.6rem;
-  }
-
-  .introHeader {
-    font-size: var(--font-large);
-  }
-
-  .address {
-    margin-right: 2.4rem;
-    font-size: 1.74rem;
-  }
-}
