+++ src/templates/Studio/IntroBlock/Block/Block.module.scss
