+++ src/templates/Studio/IntroBlock/IntroBlock.module.scss
