+++ src/templates/Studio/Recruitee/JobOffer/JobOffer.module.scss
@@ -4,9 +4,10 @@
   left: 0;
   width: 100%;
   height: 100%;
-  padding: 3.2rem 2rem 0;
+  padding: 52px 32px 0;
+  color: black;
   background: #f8f8f8;
-  z-index: 2;
+  z-index: 10;
   overflow: auto;
   -webkit-overflow-scrolling: touch;
   pointer-events: all;
@@ -18,7 +19,7 @@
   header {
     display: flex;
     position: relative;
-    margin-bottom: 0.8rem;
+    margin-bottom: 12px;
     strong {
       font-size: var(--font-small);
     }
@@ -26,11 +27,11 @@
 
   // Fields from Recruitee API
   & > div {
-    max-width: 44rem;
+    max-width: 700px;
     margin: auto;
 
     p {
-      margin-bottom: 0.6rem;
+      margin-bottom: 10px;
     }
 
     em {
@@ -40,7 +41,7 @@
     ul {
       list-style: none;
       margin: 0;
-      margin-left: 1rem;
+      margin-left: 16px;
       padding: 0;
     }
 
@@ -49,7 +50,7 @@
         content: "-";
         position: absolute;
         letter-spacing: 0.02em;
-        margin-left: -0.8rem;
+        margin-left: -12px;
       }
     }
   }
@@ -66,7 +67,7 @@
   }
 
   .closeButton {
-    width: 2.4rem;
+    width: 40px;
     font-size: var(--font-small);
     font-weight: 700;
     text-align: end;
@@ -77,9 +78,9 @@
 
   .applyButton {
     display: inline-block;
-    margin: 2rem auto 3.2rem;
+    margin: 32px auto 52px;
     border: 1px solid #000;
-    padding: 1rem;
+    padding: 16px;
     text-decoration: none;
   }
 }
@@ -87,7 +88,7 @@
 @media (min-width: 36rem) {
   .jobOffer {
     header {
-      margin-bottom: 1.2rem;
+      margin-bottom: 20px;
       strong {
         font-size: var(--font-med);
       }
@@ -96,7 +97,7 @@
     // Fields from Recruitee API
     & > div {
       p {
-        margin-bottom: 0.8rem;
+        margin-bottom: 12px;
       }
     }
