+++ src/templates/Studio/Recruitee/Recruitee.js
@@ -15,7 +15,10 @@ const Recruitee = ({ location }) => {
   useEffect(() => {
     // If fetch is not supported, we will not show any job offers
     // IE11 and lower
-    if (window.fetch) {
+    if (
+      window.fetch &&
+      (!recruiteeData.offers || !recruiteeData.offers.length)
+    ) {
       fetch('https://career.recruitee.com/api/c/23038/widget/?widget=true')
         .then(res => res.json())
         .then(json => {
@@ -37,11 +40,7 @@ const Recruitee = ({ location }) => {
           throw new Error(`Failed to fetch open job offers: ${err}`);
         });
     }
-
-    // react-hooks/exhaustive-deps about location.hash
-    // Should only check it once on render, so disabled eslint for it
-    // eslint-disable-next-line
-  }, []);
+  }, [location.hash, recruiteeData.length, recruiteeData.offers]);
 
   const handleCloseOffer = () => {
     setOpenOffer(null);
@@ -55,11 +54,10 @@ const Recruitee = ({ location }) => {
   return (
     <>
       <aside className={styles.recruitee}>
-        <p className={styles.openRoles}>Open Roles</p>
+        <p className={styles.title}>Open Positions</p>
         <ul className={styles.list}>
           {recruiteeData.offers.map(offer => (
             <li className={styles.item} key={offer.id}>
-              <p className={styles.department}>{offer.department}</p>
               <a
                 className={styles.role}
                 href={`#${offer.id}`}
