+++ src/templates/Studio/Recruitee/Recruitee.module.scss
@@ -1,52 +1,47 @@
 .recruitee {
-  margin: 0 -1.4rem 6rem;
-  padding: 0 1.6rem;
-  font-size: var(--font-small);
+  height: min-content;
+  margin: 120px 20px 120px;
 
-  @media (min-width: 36rem) {
-    font-size: var(--font-med);
-    margin-bottom: 8rem;
+  @media (min-width: 48rem) {
+    align-self: flex-end;
+    width: 50%;
+    margin: 0 20px 0px 0;
+    padding: 0 0 120px 120px;
   }
 
-  @media (min-width: 50rem) {
+  @media (min-width: 60rem) {
+    margin: 120px 20px 300px 0;
+    padding: 0 120px 0 120px;
   }
 }
 
 .list {
     margin: 0;
-    padding: 0.8rem 0 0.6rem;
     list-style: none;
+    font-size: var(--px-font-xmed);
+
+    @media (min-width: 48rem) {
+      font-size: var(--px-font-med);
+    }
+
+    @media (min-width: 60rem) {
+      font-size: var(--px-font-large);
+    }
 }
 
-.item {
-  margin-bottom: 2rem;
-
-  @media (min-width: 36rem) {
-    margin-bottom: 3rem;
-  }
-}
-
-.open-roles {
-  color: #ff0000;
-  margin-bottom: 2rem;
-
-  @media (min-width: 36rem) {
-    margin-bottom: 3rem;
-  }
-
-  @media (min-width: 50rem) {
-    flex: 1;
-    max-width: 45%;
-  }
-}
-
-.department {
-  font-size: var(--font-xs);
-  margin-bottom: 0.5rem;
+.title {
+  font-size: var(--px-font-xxs);
+  margin-bottom: 22px;
   text-transform: uppercase;
 
-  @media (min-width: 36rem) {
-    font-size: var(--font-small);
+  @media (min-width: 48rem) {
+    flex: 1;
+    font-size: var(--px-font-xs);
+    margin-bottom: 24px;
+  }
+
+  @media (min-width: 60rem) {
+    font-size: var(--px-font-small);
   }
 }
 
@@ -55,10 +50,11 @@
   background: linear-gradient(transparent, transparent),
     linear-gradient(transparent, transparent),
     linear-gradient(black, black);
-  background-size: 0.05em 1px, 0.05em 1px, 1px 0.07em;
+  background-size: 0.05em 1px, 0.05em 1px, 1px 2px;
   background-repeat: no-repeat, no-repeat, repeat-x;
   background-position: 0 87%, 100% 87%, 0 100%;
   text-decoration: none;
+  line-height: 52px;
 
   &:hover {
     opacity: 0.5;
