+++ src/templates/Studio/Studio.js
@@ -1,32 +1,49 @@
-import React from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { graphql } from 'gatsby';
+import styles from './Studio.module.scss';
 import Layout from '../../components/Layout/Layout';
-import Footer from '../../components/Footer/Footer';
-import Intro from './Intro/Intro';
-import InfoBlock from './InfoBlock/InfoBlock';
-import Impression from './Impression/Impression';
+import IntroBlock from './IntroBlock/IntroBlock';
+import ServiceList from './ServiceList/ServiceList';
 import SEO from '../../components/SEO/SEO';
-import Wonder from './Wonder/Wonder';
+import Footer from '../../components/Footer/Footer';
+import Recruitee from './Recruitee/Recruitee';
+import ImageCarousel from '../../components/ImageCarousel/ImageCarousel';
+import SkillBlock from './SkillBlock/SkillBlock';
+import useWindowSize from '../../utils/hooks/useWindowSize';
 
 export const query = graphql`
-  {
-    studioPage: markdownRemark(frontmatter: { templateKey: { eq: "Studio" } }) {
+  query Studio($templateKey: String!) {
+    studioPage: markdownRemark(
+      frontmatter: { templateKey: { eq: $templateKey } }
+    ) {
       fields {
         slug
       }
       frontmatter {
-        intro
-        infoBlock {
-          collection {
-            showIndicator
-            info
-            images {
-              caption
-              image {
-                childImageSharp {
-                  fluid(maxWidth: 1280, quality: 70) {
-                    ...GatsbyImageSharpFluid_withWebp
-                  }
+        title
+        skillset
+        introBlocks: introBlock {
+          copy
+          image {
+            childImageSharp {
+              fluid(maxWidth: 3840, quality: 90) {
+                ...GatsbyImageSharpFluid_withWebp
+              }
+            }
+          }
+          video {
+            publicURL
+          }
+        }
+        services {
+          title
+          list {
+            title
+            copy
+            image {
+              childImageSharp {
+                fluid(maxWidth: 3840, quality: 90) {
+                  ...GatsbyImageSharpFluid_withWebp
                 }
               }
             }
@@ -38,7 +55,7 @@ export const query = graphql`
           images {
             image {
               childImageSharp {
-                fluid(maxWidth: 1280, quality: 70) {
+                fluid(maxWidth: 3840, quality: 90) {
                   ...GatsbyImageSharpFluid_withWebp
                 }
               }
@@ -47,6 +64,7 @@ export const query = graphql`
         }
       }
     }
+
     indexPage: markdownRemark(frontmatter: { templateKey: { eq: "Home" } }) {
       frontmatter {
         address
@@ -57,21 +75,81 @@ export const query = graphql`
   }
 `;
 
-// eslint-disable-next-line react/display-name
-export default ({ location, data: { indexPage, studioPage } }) => {
+const mediumBreakpoint = 960; // BP of 60rem
+
+const Studio = ({
+  data: {
+    studioPage: { fields, frontmatter },
+    indexPage,
+  },
+  location,
+}) => {
+  const introRef = useRef();
+  const [themeClass, setThemeClass] = useState();
+  const { width } = useWindowSize();
+
+  useEffect(() => {
+    if (window.IntersectionObserver) {
+      setThemeClass(styles.darkTheme);
+
+      const options = {
+        rootMargin: '-120px 0px 0px 0px',
+      };
+
+      const cb = entries => {
+        entries.forEach(entry => {
+          if (entry.isIntersecting) {
+            setThemeClass(styles.darkTheme);
+          } else {
+            setThemeClass('');
+          }
+        });
+      };
+
+      const observer = new IntersectionObserver(cb, options);
+      observer.observe(introRef.current);
+    }
+  }, []);
+
   return (
     <Layout>
-      <SEO title="Studio" pathName={studioPage.fields.slug} />
-      <Wonder />
-      <Intro
-        data={{ ...indexPage.frontmatter, ...studioPage.frontmatter }}
-        location={location}
-      />
-      {studioPage.frontmatter.infoBlock.map(({ collection }, index) => (
-        <InfoBlock key={index} collection={collection} />
-      ))}
-      <Impression data={studioPage.frontmatter.studioImpression} />
-      <Footer {...indexPage.frontmatter} />
+      <div
+        tabIndex="-1" // Makes it scrollable with keyboard
+        className={`${styles.wrapper} ${
+          width > mediumBreakpoint ? themeClass : ''
+        }`}
+      >
+        <SEO title="Studio" pathName={fields.slug} />
+        <IntroBlock
+          title={frontmatter.title}
+          intros={frontmatter.introBlocks}
+          ref={introRef}
+        />
+
+        <ServiceList
+          services={frontmatter.services.list}
+          headerTitle={frontmatter.services.title}
+        />
+
+        <SkillBlock
+          skillset={frontmatter.skillset}
+          email={indexPage.frontmatter.email}
+        />
+
+        <div className={styles.jobsImpressionBlock}>
+          <Recruitee location={location} />
+          <ImageCarousel
+            className={styles.carouselWrapper}
+            images={frontmatter.studioImpression.images}
+            showIndicator={frontmatter.studioImpression.showIndicator}
+            title={frontmatter.studioImpression.title}
+          />
+        </div>
+
+        <Footer {...indexPage.frontmatter} />
+      </div>
     </Layout>
   );
 };
+
+export default Studio;
