--- src/templates/Studio/Wonder/Wonder.module.scss
@@ -1,16 +1,0 @@
-.canvas {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100vh;
-  width: 100vw;
-  z-index: 0;
-  pointer-events: none;
-  opacity: 0.01;
-  transition: opacity 4000ms ease-in 500ms;
-  outline: none;
-
-  &.is-visible {
-    opacity: 1;
-  }
-}
