--- src/templates/Studio2/Recruitee/JobOffer/JobOffer.module.scss
@@ -1,108 +1,0 @@
-.jobOffer {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  padding: 52px 32px 0;
-  color: black;
-  background: #f8f8f8;
-  z-index: 10;
-  overflow: auto;
-  -webkit-overflow-scrolling: touch;
-  pointer-events: all;
-
-  & * {
-    font-weight: normal;
-  }
-
-  header {
-    display: flex;
-    position: relative;
-    margin-bottom: 12px;
-    strong {
-      font-size: var(--font-small);
-    }
-  }
-
-  // Fields from Recruitee API
-  & > div {
-    max-width: 700px;
-    margin: auto;
-
-    p {
-      margin-bottom: 10px;
-    }
-
-    em {
-      font-style: normal;
-    }
-
-    ul {
-      list-style: none;
-      margin: 0;
-      margin-left: 16px;
-      padding: 0;
-    }
-
-    li {
-      &::before {
-        content: "-";
-        position: absolute;
-        letter-spacing: 0.02em;
-        margin-left: -12px;
-      }
-    }
-  }
-
-  button {
-    border: none;
-    border-radius: 0;
-    background: none;
-  }
-
-  .closeButton,
-  .applyButton {
-    outline: none;
-  }
-
-  .closeButton {
-    width: 40px;
-    font-size: var(--font-small);
-    font-weight: 700;
-    text-align: end;
-    margin: 0;
-    padding: 0;
-    margin-left: auto;
-  }
-
-  .applyButton {
-    display: inline-block;
-    margin: 32px auto 52px;
-    border: 1px solid #000;
-    padding: 16px;
-    text-decoration: none;
-  }
-}
-
-@media (min-width: 36rem) {
-  .jobOffer {
-    header {
-      margin-bottom: 20px;
-      strong {
-        font-size: var(--font-med);
-      }
-    }
-
-    // Fields from Recruitee API
-    & > div {
-      p {
-        margin-bottom: 12px;
-      }
-    }
-
-    .closeButton {
-      font-size: var(--font-med);
-    }
-  }
-}
