--- src/templates/Studio2/Recruitee/Recruitee.module.scss
@@ -1,63 +1,0 @@
-.recruitee {
-  height: min-content;
-  margin: 120px 20px 120px;
-
-  @media (min-width: 48rem) {
-    align-self: flex-end;
-    width: 50%;
-    margin: 0 20px 0px 0;
-    padding: 0 0 120px 120px;
-  }
-
-  @media (min-width: 60rem) {
-    margin: 120px 20px 300px 0;
-    padding: 0 120px 0 120px;
-  }
-}
-
-.list {
-    margin: 0;
-    list-style: none;
-    font-size: var(--px-font-xmed);
-
-    @media (min-width: 48rem) {
-      font-size: var(--px-font-med);
-    }
-
-    @media (min-width: 60rem) {
-      font-size: var(--px-font-large);
-    }
-}
-
-.title {
-  font-size: var(--px-font-xxs);
-  margin-bottom: 22px;
-  text-transform: uppercase;
-
-  @media (min-width: 48rem) {
-    flex: 1;
-    font-size: var(--px-font-xs);
-    margin-bottom: 24px;
-  }
-
-  @media (min-width: 60rem) {
-    font-size: var(--px-font-small);
-  }
-}
-
-.role {
-  transition: opacity 0.2s cubic-bezier(0.445, 0.05, 0.55, 0.95);
-  background: linear-gradient(transparent, transparent),
-    linear-gradient(transparent, transparent),
-    linear-gradient(black, black);
-  background-size: 0.05em 1px, 0.05em 1px, 1px 2px;
-  background-repeat: no-repeat, no-repeat, repeat-x;
-  background-position: 0 87%, 100% 87%, 0 100%;
-  text-decoration: none;
-  line-height: 52px;
-
-  &:hover {
-    opacity: 0.5;
-    transition: opacity 0s cubic-bezier(0.445, 0.05, 0.55, 0.95) !important;
-  }
-}
